Fix redundant cast warning in FlightTransport
Remove unnecessary cast to int in client index calculation as
modulo operation already returns int type.

+++ plugins/arrow-flight-rpc/src/main/java/org/opensearch/arrow/flight/transport/FlightTransport.java
@@ -354,7 +354,7 @@ class FlightTransport extends TcpTransport {
         });
         
         // Round-robin client selection based on channel counter
-        int clientIndex = (int) (GLOBAL_CHANNEL_COUNTER.incrementAndGet() % holder.flightClients().size());
+        int clientIndex = GLOBAL_CHANNEL_COUNTER.incrementAndGet() % holder.flightClients().size();
         
         FlightClientChannel channel = new FlightClientChannel(
             boundAddress,
